Move calendar to local.py
