+++ homeassistant/components/calendar/__init__.py
@@ -8,9 +8,6 @@ import asyncio
 import logging
 from datetime import timedelta
 import re
-import uuid
-
-import voluptuous as vol
 
 from homeassistant.components.google import (
     CONF_OFFSET, CONF_DEVICE_ID, CONF_NAME)
@@ -21,12 +18,6 @@ from homeassistant.helpers.entity import Entity, generate_entity_id
 from homeassistant.helpers.entity_component import EntityComponent
 from homeassistant.helpers.template import DATE_STR_FORMAT
 from homeassistant.util import dt
-from homeassistant.components import http
-from homeassistant.core import callback
-import homeassistant.helpers.config_validation as cv
-from homeassistant.util.json import load_json, save_json
-
-
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -36,61 +27,12 @@ ENTITY_ID_FORMAT = DOMAIN + '.{}'
 
 SCAN_INTERVAL = timedelta(seconds=60)
 
-EVENT = 'calendar_updated'
-
-EVENT_SCHEMA = vol.Schema({
-    vol.Required('title'): cv.string,
-    vol.Required('start'): cv.string,
-    vol.Optional('end'): cv.string,
-    vol.Optional('url'): cv.string,
-    vol.Optional('color'): cv.string,
-    vol.Optional('all_day'): cv.boolean,
-    vol.Optional('description'): cv.string,
-})
-
-PERSISTENCE = '.calendar.json'
-
 
 @asyncio.coroutine
 def async_setup(hass, config):
     """Track states and offer events for calendars."""
-    data = hass.data[DOMAIN] = CalendarData(hass)
-    yield from data.async_load()
-
     component = EntityComponent(
         _LOGGER, DOMAIN, hass, SCAN_INTERVAL, DOMAIN)
-
-    hass.http.register_view(CalendarView)
-
-    async def create_event(call):
-        title = call.data.get("title")
-        start = call.data.get("start")
-        end = call.data.get("end")
-        url = call.data.get("url")
-        color = call.data.get("color")
-        all_day = call.data.get("all_day", False)
-        description = call.data.get("description")
-
-        hass.data[DOMAIN].async_add(title, start, end, url,
-                                    color, all_day, description)
-        hass.bus.async_fire(EVENT)
-
-    hass.services.async_register(DOMAIN, 'create_event', create_event,
-                                 schema=EVENT_SCHEMA)
-
-    async def delete_event(call):
-        title = call.data.get("title")
-        start = call.data.get("start")
-
-        hass.data[DOMAIN].async_delete(title, start)
-        hass.bus.async_fire(EVENT)
-
-    hass.services.async_register(DOMAIN, 'delete_event', delete_event,
-                                 schema=EVENT_SCHEMA)
-
-    yield from hass.components.frontend.async_register_built_in_panel(
-        'calendar', 'Calendar', 'mdi:calendar')
-
 
     yield from component.async_setup(config)
     return True
@@ -241,80 +183,3 @@ class CalendarEventDevice(Entity):
         self._cal_data['start'] = start
         self._cal_data['end'] = end
         self._cal_data['all_day'] = 'date' in self.data.event['start']
-
-
-class CalendarData:
-    """Class to hold scheduler data."""
-
-    def __init__(self, hass):
-        """Initialize the scheduler."""
-        self.hass = hass
-        self.items = []
-
-    def _get_item(self, title, start):
-        # Check if the object exists
-        for item in self.items:
-            if title == item['title'] and start == item['start']:
-                # Item already created
-                return item
-        return None
-
-    @callback
-    def async_add(self, title, start,
-                  end=None, url=None, color=None,
-                  all_day=False, description=None):
-        """Add a scheduler item."""
-        # Check if the object exists
-        if self._get_item(title, start) is not None:
-            return
-        # Create new item
-        item = {
-            'title': title,
-            'start': start,
-            'all_day': all_day,
-            'id': uuid.uuid4().hex,
-        }
-        for key in ('end', 'url', 'color', 'description'):
-            value = locals().get(key)
-            if value is not None:
-                item[key] = value
-        self.items.append(item)
-        self.hass.async_add_job(self.save)
-        return item
-
-    @callback
-    def async_update(self, item_id, info):
-        """Update a scheduler item."""
-        raise NotImplementedError
-
-    @callback
-    def async_delete(self, title, date):
-        """Clear completed items."""
-        # TODO
-        self.items = [itm for itm in self.items if not itm['complete']]
-        self.hass.async_add_job(self.save)
-
-    @asyncio.coroutine
-    def async_load(self):
-        """Load items."""
-        def load():
-            """Load the items synchronously."""
-            return load_json(self.hass.config.path(PERSISTENCE), default=[])
-
-        self.items = yield from self.hass.async_add_job(load)
-
-    def save(self):
-        """Save the items."""
-        save_json(self.hass.config.path(PERSISTENCE), self.items)
-
-
-class CalendarView(http.HomeAssistantView):
-    """View to retrieve calendar content."""
-
-    url = '/api/calendar'
-    name = "api:calendar"
-
-    @callback
-    def get(self, request):
-        """Retrieve calendar items."""
-        return self.json(request.app['hass'].data[DOMAIN].items)
